requirements.txt
langchain-huggingface
huggingface-hub
langchain-core
langchain
python-dotenv
streamlit